FOUR-80 Fix: journals 수정에서 default-image인 경우 처리 추가

+++ src/main/java/com/hallym/booker/service/JournalsService.java
@@ -62,7 +62,9 @@ public class JournalsService {
         String imageName = journals.getJimageName();
         String imageUrl = journals.getJimageUrl();
         if(journalsEditRequest.getFile() != null) {
-            s3Service.delete(journals.getJimageName());
+            if(!journals.getJimageName().equals("default-profile.png")) {
+                s3Service.delete(journals.getJimageName());
+            }
             S3ResponseUploadEntity journal = s3Service.upload(journalsEditRequest.getFile(), "journal");
 
             imageName = journal.getImageName();
